+++ plugins/main/public/components/overview/vulnerabilities/dashboards/overview/dashboard.tsx
@@ -3,7 +3,7 @@ import { getPlugins } from '../../../../../kibana-services';
 import { ViewMode } from '../../../../../../../../src/plugins/embeddable/public';
 import { getDashboardPanels } from './dashboard_panels';
 import { I18nProvider } from '@osd/i18n/react';
-import useSearchBarConfiguration from '../../search_bar/use_search_bar_configuration';
+import useSearchBar from '../../../../common/search-bar/use-search-bar';
 import { getDashboardFilters } from './dashboard_panels_filters';
 import './vulnerability_detector_filters.scss';
 import { getKPIsPanel } from './dashboard_panels_kpis';
@@ -27,7 +27,7 @@ const DashboardVulsComponent: React.FC = () => {
   const appConfig = useAppConfig();
   const VULNERABILITIES_INDEX_PATTERN_ID =
     appConfig.data['vulnerabilities.pattern'];
-  const { searchBarProps } = useSearchBarConfiguration({
+  const { searchBarProps } = useSearchBar({
     defaultIndexPatternID: VULNERABILITIES_INDEX_PATTERN_ID,
   });
   const {
@@ -62,11 +62,17 @@ const DashboardVulsComponent: React.FC = () => {
           ) : null}
           {!isLoadingSearchbar &&
           !isLoading &&
-          (isError || resultIndexData?.hits?.total === 0) ? (
+          (isError ||
+            !resultIndexData ||
+            resultIndexData?.hits?.total === 0) ? (
             <DiscoverNoResults message={error?.message} />
           ) : null}
-          {!isLoadingSearchbar && !isLoading && isSuccess ? (
-            <>
+          {!isLoadingSearchbar &&
+          !isLoading &&
+          isSuccess &&
+          resultIndexData &&
+          resultIndexData?.hits?.total !== 0 ? (
+            <div className='vulnerability-dashboard-responsive'>
               <div className='vulnerability-dashboard-filters-wrapper'>
                 <DashboardByRenderer
                   input={{
@@ -138,7 +144,7 @@ const DashboardVulsComponent: React.FC = () => {
                   hidePanelTitles: false,
                 }}
               />
-            </>
+            </div>
           ) : null}
         </>
       </I18nProvider>