+++ plugins/main/public/components/overview/vulnerabilities/search_bar/index.ts
@@ -1,1 +1,0 @@
-// searchbar index