+++ plugins/main/public/kibana-integrations/kibana-vis.js
@@ -33,7 +33,6 @@ import {
   EuiIcon,
   EuiFlexItem,
   EuiFlexGroup,
-  EuiEmptyPrompt
 } from '@elastic/eui';
 import {
   getAngularModule,
@@ -45,7 +44,6 @@ import {
   getOverlays,
   getPlugins,
 } from '../kibana-services';
-import { KnownFields } from '../utils/known-fields';
 import { union } from 'lodash';
 import { getFilterWithAuthorizedAgents } from '../react-services/filter-authorization-agents';
 import { AUTHORIZED_AGENTS } from '../../common/constants';
@@ -86,7 +84,9 @@ class KibanaVis extends Component {
       ...services,
       ...{ visualizationTypes: getVisualizationsPlugin() },
     };
-    this.savedObjectLoaderVisualize = createSavedVisLoader(servicesForVisualizations);
+    this.savedObjectLoaderVisualize = createSavedVisLoader(
+      servicesForVisualizations,
+    );
     this.visID = this.props.visID;
     this.tab = this.props.tab;
   }
@@ -114,7 +114,7 @@ class KibanaVis extends Component {
       // componentWillUnmount.
       // In the renderComplete() the value of the $rootScope
       // is changed, which affects the entire application.
-      // 
+      //
       // Related issue:
       // https://github.com/wazuh/wazuh-dashboard-plugins/issues/4158
       if (this.deadField) {
@@ -149,13 +149,14 @@ class KibanaVis extends Component {
           data.value &&
           data.value.visData &&
           data.value.visData.rows &&
-          this.props.state[this.visID] !== data.value.visData.rows['0']['col-0-1']
+          this.props.state[this.visID] !==
+            data.value.visData.rows['0']['col-0-1']
         ) {
           store.dispatch(
             this.updateMetric({
               name: this.visID,
               value: data.value.visData.rows['0']['col-0-1'],
-            })
+            }),
           );
         }
         // This check if data.value.visData.tables exists and dispatch that value as stat
@@ -169,13 +170,14 @@ class KibanaVis extends Component {
           data.value.visData.tables['0'] &&
           data.value.visData.tables['0'].rows &&
           data.value.visData.tables['0'].rows['0'] &&
-          this.props.state[this.visID] !== data.value.visData.tables['0'].rows['0']['col-0-2']
+          this.props.state[this.visID] !==
+            data.value.visData.tables['0'].rows['0']['col-0-2']
         ) {
           store.dispatch(
             this.updateMetric({
               name: this.visID,
               value: data.value.visData.tables['0'].rows['0']['col-0-2'],
-            })
+            }),
           );
         }
       }
@@ -195,17 +197,18 @@ class KibanaVis extends Component {
     }
   };
 
-  setSearchSource = (discoverList) => {
+  setSearchSource = discoverList => {
     try {
       const isCluster = this.visID.includes('Cluster');
       if (isCluster) {
         // Checks for cluster.name or cluster.node filter existence
         const monitoringFilter = discoverList[1].filter(
-          (item) =>
+          item =>
             item &&
             item.meta &&
             item.meta.key &&
-            (item.meta.key.includes('cluster.name') || item.meta.key.includes('cluster.node'))
+            (item.meta.key.includes('cluster.name') ||
+              item.meta.key.includes('cluster.node')),
         );
 
         // Applying specific filter to cluster monitoring vis
@@ -218,12 +221,15 @@ class KibanaVis extends Component {
     }
   };
 
-  myRender = async (raw) => {
+  myRender = async raw => {
     const timefilter = getDataPlugin().query.timefilter.timefilter;
     try {
       const discoverList = this.discoverPendingUpdates.getList();
-      const isAgentStatus = this.visID === 'Wazuh-App-Overview-General-Agents-status';
-      const timeFilterSeconds = this.calculateTimeFilterSeconds(timefilter.getTime());
+      const isAgentStatus =
+        this.visID === 'Wazuh-App-Overview-General-Agents-status';
+      const timeFilterSeconds = this.calculateTimeFilterSeconds(
+        timefilter.getTime(),
+      );
       const timeRange =
         isAgentStatus && timeFilterSeconds < 900
           ? { from: 'now-15m', to: 'now', mode: 'quick' }
@@ -243,8 +249,13 @@ class KibanaVis extends Component {
                     {
                       term:
                         AppState.getClusterInfo().status === 'enabled'
-                          ? { 'cluster.name': AppState.getClusterInfo().cluster }
-                          : { 'manager.keyword': AppState.getClusterInfo().manager },
+                          ? {
+                              'cluster.name': AppState.getClusterInfo().cluster,
+                            }
+                          : {
+                              'manager.keyword':
+                                AppState.getClusterInfo().manager,
+                            },
                     },
                   ],
                 },
@@ -257,19 +268,29 @@ class KibanaVis extends Component {
         : discoverList[1] || [];
       const query = !isAgentStatus ? discoverList[0] : {};
 
-      const rawVis = raw ? raw.filter((item) => item && item.id === this.visID) : [];
+      const rawVis = raw
+        ? raw.filter(item => item && item.id === this.visID)
+        : [];
 
       if (rawVis.length && discoverList.length) {
         let vizPattern;
         try {
-          vizPattern = JSON.parse(rawVis[0].attributes.kibanaSavedObjectMeta.searchSourceJSON)
-            .index;
+          vizPattern = JSON.parse(
+            rawVis[0].attributes.kibanaSavedObjectMeta.searchSourceJSON,
+          ).index;
         } catch (ex) {
           console.warn(`plugin platform-vis exception: ${ex.message || ex}`);
         }
 
-        if (!filters.find((filter) => filter.meta.controlledBy === AUTHORIZED_AGENTS)) {
-          const agentsFilters = getFilterWithAuthorizedAgents(this.props.allowedAgents, vizPattern);
+        if (
+          !filters.find(
+            filter => filter.meta.controlledBy === AUTHORIZED_AGENTS,
+          )
+        ) {
+          const agentsFilters = getFilterWithAuthorizedAgents(
+            this.props.allowedAgents,
+            vizPattern,
+          );
           filters = agentsFilters ? union(filters, [agentsFilters]) : filters;
         }
 
@@ -284,16 +305,21 @@ class KibanaVis extends Component {
         if (!this.visualization && !this.rendered && !this.renderInProgress) {
           // There's no visualization object -> create it with proper filters
           this.renderInProgress = true;
-          this.visualization = await this.savedObjectLoaderVisualize.get(this.visID, rawVis[0]);
-          this.visualization.searchSource = await getDataPlugin().search.searchSource.create();
+          this.visualization = await this.savedObjectLoaderVisualize.get(
+            this.visID,
+            rawVis[0],
+          );
+          this.visualization.searchSource =
+            await getDataPlugin().search.searchSource.create();
           // Visualization doesn't need the "_source"
           this.visualization.searchSource.setField('source', false);
           // Visualization doesn't need "hits"
           this.visualization.searchSource.setField('size', 0);
-          const visState = await getVisualizationsPlugin().convertToSerializedVis(
-            this.visualization
-          );
-          
+          const visState =
+            await getVisualizationsPlugin().convertToSerializedVis(
+              this.visualization,
+            );
+
           // In Kibana 7.10.2, there is a bug when creating the visualization with `createVis` method of the Visualization plugin that doesn't pass the `visState` parameter to the `Vis` class constructor.
           // This does the `.params`, `.uiState` and `.id` properties of the visualization are not set correctly in the `Vis` class constructor. This bug causes
           // that the visualization, for example, doesn't use the defined colors in the `.uiStateJSON` property.
@@ -305,16 +331,17 @@ class KibanaVis extends Component {
           // as is done in the `Vis` class constructor https://github.com/elastic/kibana/blob/v7.10.2/src/plugins/visualizations/public/vis.ts#L99-L104
           const vis = await getVisualizationsPlugin().createVis(
             this.visualization.visState.type,
-            visState
+            visState,
           );
           vis.params = vis.getParams(visState.params);
           vis.uiState = new PersistedState(visState.uiState);
           vis.id = visState.id;
-          
-          this.visHandler = await getVisualizationsPlugin().__LEGACY.createVisEmbeddableFromObject(
-            vis,
-            visInput
-          );
+
+          this.visHandler =
+            await getVisualizationsPlugin().__LEGACY.createVisEmbeddableFromObject(
+              vis,
+              visInput,
+            );
           await this.visHandler.render($(`[id="${this.visID}"]`)[0]);
           this.visHandler.handler.data$.subscribe(this.renderComplete());
           this.visHandlers.addItem(this.visHandler);
@@ -331,10 +358,15 @@ class KibanaVis extends Component {
           this.visHandler.updateInput(visInput);
           this.setSearchSource(discoverList);
         }
-        if (this.state.visRefreshingIndex) this.setState({ visRefreshingIndex: false });
+        if (this.state.visRefreshingIndex)
+          this.setState({ visRefreshingIndex: false });
       }
     } catch (error) {
-      if (((error || {}).message || '').includes('not locate that index-pattern-field')) {
+      if (
+        ((error || {}).message || '').includes(
+          'not locate that index-pattern-field',
+        )
+      ) {
         if (this.deadField) {
           this.tabVisualizations.addDeadVis();
           return this.renderComplete();
@@ -349,11 +381,7 @@ class KibanaVis extends Component {
           this.renderInProgress = false;
           this.rendered = false;
 
-          // if there's a field name it looks for known fields structures
-          const foundField =
-            match[1] && KnownFields.find((field) => field.name === match[1].trim());
-
-          await this.props.refreshKnownFields(foundField);
+          await this.props.refreshKnownFields();
         }
         this.renderInProgress = false;
         return this.myRender(raw);
@@ -387,8 +415,10 @@ class KibanaVis extends Component {
     this.loadedVisualizations.addItem(true);
 
     const currentLoaded = this.loadedVisualizations.getList().length;
-    const deadVis = this.props.tab === 'ciscat' ? 0 : this.tabVisualizations.getDeadVis();
-    const totalTabVis = this.tabVisualizations.getItem(this.props.tab) - deadVis;
+    const deadVis =
+      this.props.tab === 'ciscat' ? 0 : this.tabVisualizations.getDeadVis();
+    const totalTabVis =
+      this.tabVisualizations.getItem(this.props.tab) - deadVis;
     this.$rootScope.loadingStatus = 'Fetching data...';
 
     if (totalTabVis < 1) {
@@ -398,7 +428,9 @@ class KibanaVis extends Component {
       if (currentCompleted >= 100) {
         this.$rootScope.rendered = 'true';
         if (visId.includes('AWS-geo')) {
-          const canvas = $('.visChart.leaflet-container .leaflet-control-zoom-in');
+          const canvas = $(
+            '.visChart.leaflet-container .leaflet-control-zoom-in',
+          );
           setTimeout(() => {
             if (!this.mapClicked) {
               this.mapClicked = true;
@@ -410,23 +442,25 @@ class KibanaVis extends Component {
         this.$rootScope.rendered = 'false';
       }
     }
-    if (typeof this.props.onRenderComplete === 'function') this.props.onRenderComplete();
+    if (typeof this.props.onRenderComplete === 'function')
+      this.props.onRenderComplete();
   };
 
   showDateRangePicker = () => {
-    return !this.deadField && !this.state.visRefreshingIndex && this.visID === 'Wazuh-App-Overview-General-Agents-status'
-  }
+    return (
+      !this.deadField &&
+      !this.state.visRefreshingIndex &&
+      this.visID === 'Wazuh-App-Overview-General-Agents-status'
+    );
+  };
 
   DateRangePickerComponent = () => {
     return (
-      <EuiFlexItem className="agents-evolutions-dpicker">
-        <WzDatePicker
-          condensed={true} 
-          onTimeChange={() => {}} 
-        />
+      <EuiFlexItem className='agents-evolutions-dpicker'>
+        <WzDatePicker condensed={true} onTimeChange={() => {}} />
       </EuiFlexItem>
-    )
-  }
+    );
+  };
 
   render() {
     const isLoading = this.props.resultState === 'loading';
@@ -443,7 +477,7 @@ class KibanaVis extends Component {
             <EuiFlexGroup style={{ placeItems: 'center' }}>
               <EuiFlexItem></EuiFlexItem>
               <EuiFlexItem grow={false}>
-                <EuiLoadingSpinner size="xl" />
+                <EuiLoadingSpinner size='xl' />
               </EuiFlexItem>
               <EuiFlexItem grow={false}>Refreshing Index Pattern.</EuiFlexItem>
 
@@ -452,77 +486,90 @@ class KibanaVis extends Component {
           </div>
           <div
             style={{
-              display: isLoading && !this.state.visRefreshingIndex ? 'block' : 'none',
+              display:
+                isLoading && !this.state.visRefreshingIndex ? 'block' : 'none',
               textAlign: 'center',
               paddingTop: 100,
             }}
           >
-            <EuiLoadingChart size="xl" />
+            <EuiLoadingChart size='xl' />
           </div>
           <div
             style={{
               display:
-                this.deadField && !isLoading && !this.state.visRefreshingIndex ? 'block' : 'none',
+                this.deadField && !isLoading && !this.state.visRefreshingIndex
+                  ? 'block'
+                  : 'none',
               textAlign: 'center',
               paddingTop: 100,
             }}
           >
             No results found &nbsp;
             <EuiToolTip
-              position="top"
+              position='top'
               content={
                 <span>
-                  No alerts were found with the field: <strong>{this.deadField}</strong>
+                  No alerts were found with the field:{' '}
+                  <strong>{this.deadField}</strong>
                 </span>
               }
             >
-              <EuiIcon type="iInCircle" />
+              <EuiIcon type='iInCircle' />
             </EuiToolTip>
           </div>
-          {   
-            !this.isLoading && this.showDateRangePicker() &&
-            this.DateRangePickerComponent()
-          }
+          {!this.isLoading &&
+            this.showDateRangePicker() &&
+            this.DateRangePickerComponent()}
           <div
             id={this.visID}
             vis-id={this.visID}
-            style={{ 
-              display: isLoading ? 'none' : 'block', 
+            style={{
+              display: isLoading ? 'none' : 'block',
               height: '100%',
-              paddingTop: '2%' 
+              paddingTop: '2%',
             }}
           ></div>
-          <div style={{
-                position: 'absolute',
-                top: '50%',
-                left: '50%',
-                transform: 'translate(-50%, -50%)',
-              }}>
-                {(isLoading && <div><EuiLoadingChart size="xl" /></div>)
-                  || (this.deadField && !isLoading && !this.state.visRefreshingIndex && (
-                    <div>
-                      No results found &nbsp;
-                      <EuiToolTip
-                        position="top"
-                        content={
-                          <span>
-                            No alerts were found with the field: <strong>{this.deadField}</strong>
-                          </span>
-                        }
-                      >
-                        <EuiIcon type="iInCircle" />
-                      </EuiToolTip>
-                    </div>
-                  ))
-                  || (this.state.visRefreshingIndex && (
-                    <EuiFlexGroup justifyContent="center" alignItems="center">
-                      <EuiFlexItem grow={false}>
-                        <EuiLoadingSpinner size="xl" />
-                      </EuiFlexItem>
-                      <EuiFlexItem grow={false}>Refreshing Index Pattern.</EuiFlexItem>
-                    </EuiFlexGroup>
-                  ))
-                }
+          <div
+            style={{
+              position: 'absolute',
+              top: '50%',
+              left: '50%',
+              transform: 'translate(-50%, -50%)',
+            }}
+          >
+            {(isLoading && (
+              <div>
+                <EuiLoadingChart size='xl' />
+              </div>
+            )) ||
+              (this.deadField &&
+                !isLoading &&
+                !this.state.visRefreshingIndex && (
+                  <div>
+                    No results found &nbsp;
+                    <EuiToolTip
+                      position='top'
+                      content={
+                        <span>
+                          No alerts were found with the field:{' '}
+                          <strong>{this.deadField}</strong>
+                        </span>
+                      }
+                    >
+                      <EuiIcon type='iInCircle' />
+                    </EuiToolTip>
+                  </div>
+                )) ||
+              (this.state.visRefreshingIndex && (
+                <EuiFlexGroup justifyContent='center' alignItems='center'>
+                  <EuiFlexItem grow={false}>
+                    <EuiLoadingSpinner size='xl' />
+                  </EuiFlexItem>
+                  <EuiFlexItem grow={false}>
+                    Refreshing Index Pattern.
+                  </EuiFlexItem>
+                </EuiFlexGroup>
+              ))}
           </div>
         </span>
       )
@@ -530,7 +577,7 @@ class KibanaVis extends Component {
   }
 }
 
-const mapStateToProps = (state) => {
+const mapStateToProps = state => {
   return {
     state: state.visualizationsReducers,
     allowedAgents: state.appStateReducers.allowedAgents,